requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
requests==2.32.3
urllib3==2.2.2
pycryptodome==3.20.0
protobuf>=5.28.0